pyproject.toml
[project]
name = "whoop-coach"
version = "0.1.0"
description = "WHOOP × Telegram Coach — персональный тренировочный ассистент"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "python-telegram-bot[webhooks]>=20",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "aiosqlite>=0.19",
    "pydantic-settings>=2",
    "httpx>=0.26",
    "cryptography>=42",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whoop_coach"]

[tool.ruff]
target-version = "py311"
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
